Pygroup-WebKanban/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
django-cors-headers==3.11.0
django==4.0.3
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
